requirements.txt
fastapi 
uvicorn
pydantic-settings
SQLAlchemy>=2.0
alembic>=1.14
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart